refactor(netstandard-library): move and rename configuration classes.
